requirements.txt
numpy>=1.26.0
matplotlib>=3.8.0
pydantic>=2.5.0
PyYAML>=6.0.1
pandas>=2.1.0
scipy>=1.11.0
rich>=13.7.0
click>=8.1.7
